pyproject.toml
[project]
name = "url-shortener"
version = "0.1.0"
description = ""
authors = [{ name = "Nikolay Chechulin", email = "nchechulin@protonmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = ["fastapi (>=0.115.12,<0.116.0)", "uvicorn[standard] (>=0.34.0,<0.35.0)", "sqlalchemy (>=2.0.39,<3.0.0)", "psycopg2-binary (>=2.9.10,<3.0.0)", "redis (>=5.2.1,<6.0.0)", "python-jose (>=3.4.0,<4.0.0)", "passlib[bcrypt] (>=1.7.4,<2.0.0)", "pydantic[email] (>=2.10.6,<3.0.0)", "python-multipart (>=0.0.20,<0.0.21)", "alembic (>=1.15.1,<2.0.0)", "sqlalchemy-utils (>=0.41.2,<0.42.0)"]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.2"
pre-commit = "^4.2.0"
poetry-plugin-export = ">=1.9"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
